+++ src/main/kotlin/icu/windea/pls/lang/model/ParadoxGameType.kt
@@ -40,4 +40,6 @@ enum class ParadoxGameType(
 
 val ParadoxGameType?.id get() = this?.id ?: "core"
 
+val ParadoxGameType?.linkToken get() = if(this == null) "" else "${id}:"
+
 fun ParadoxGameType?.orDefault() = this ?: getSettings().defaultGameType