+++ src/main/kotlin/icu/windea/pls/lang/parameter/impl/ParadoxDefinitionParameterSupport.kt
@@ -10,6 +10,7 @@ import icu.windea.pls.core.*
 import icu.windea.pls.core.psi.*
 import icu.windea.pls.core.search.*
 import icu.windea.pls.core.search.selector.chained.*
+import icu.windea.pls.lang.model.*
 import icu.windea.pls.lang.parameter.*
 import icu.windea.pls.script.psi.*
 
@@ -98,7 +99,7 @@ class ParadoxDefinitionParameterSupport : ParadoxParameterSupport {
         append(": ")
         
         val type = definitionType.first()
-        val typeLink = "${gameType.id}/types/${type}"
+        val typeLink = "${gameType.linkToken}types/${type}"
         appendCwtLink(typeLink, type)
         for((index, t) in definitionType.withIndex()) {
             if(index == 0) continue